Project_Chatbot/requirements.txt
langchain_openai
langchain_core
python_dotenv
streamlit
langchain_community
langserve
fastapi
unicorn
sse_starlette
bs4
pypdf
langchain_google_vertexai